Move kv files to subdir
