+++ src/cpp_wrapper/BppTree.h
@@ -10,7 +10,7 @@
 #define PROJECT_BPPTREE_H
 
 #include "Dictionary.h"
-#include "../kv_system.h"
+#include "../key_value/kv_system.h"
 #include "../dictionary/bpp_tree/bpp_tree_handler.h"
 
 template<typename K, typename V>
