+++ src/cpp_wrapper/FlatFile.h
@@ -10,7 +10,7 @@
 #define PROJECT_FLATFILE_H
 
 #include "Dictionary.h"
-#include "../kv_system.h"
+#include "../key_value/kv_system.h"
 #include "../dictionary/flat_file/flat_file_dictionary_handler.h"
 
 template<typename K, typename V>
