+++ src/cpp_wrapper/OpenAddressFileHash.h
@@ -11,7 +11,7 @@
 #define PROJECT_OPENADDRESSFILEHASH_H
 
 #include "Dictionary.h"
-#include "../kv_system.h"
+#include "../key_value/kv_system.h"
 #include "../dictionary/open_address_file_hash/open_address_file_hash_dictionary_handler.h"
 
 template<typename K, typename V>
