+++ src/cpp_wrapper/SkipList.h
@@ -10,7 +10,7 @@
 #define PROJECT_SKIPLIST_H
 
 #include "Dictionary.h"
-#include "../kv_system.h"
+#include "../key_value/kv_system.h"
 #include "../dictionary/skip_list/skip_list_handler.h"
 
 template<typename K, typename V>
