+++ src/dictionary/bpp_tree/bpp_tree.h
@@ -10,7 +10,7 @@ extern "C" {
 #include <stdarg.h>
 #include <string.h>
 #include <stdint.h>
-#include "./../../kv_system.h"
+#include "../../key_value/kv_system.h"
 #include "./../dictionary.h"
 #include "./../../file/ion_file.h"
