+++ src/dictionary/bpp_tree/bpp_tree_handler.c
@@ -7,8 +7,8 @@
 /******************************************************************************/
 
 #include "bpp_tree_handler.h"
-#include "./../../kv_system.h"
-#include "./../../kv_io.h"
+#include "../../key_value/kv_system.h"
+#include "../../key_value/kv_io.h"
 
 void
 bpptree_get_addr_filename(
