+++ src/dictionary/bpp_tree/bpp_tree_handler.h
@@ -15,7 +15,7 @@ extern "C" {
 
 #include "../dictionary_types.h"
 #include "./../dictionary.h"
-#include "./../../kv_system.h"
+#include "../../key_value/kv_system.h"
 #include "../../file/linked_file_bag.h"
 #include "bpp_tree.h"
