+++ src/dictionary/flat_file/flat_file.c
@@ -6,7 +6,7 @@
  */
 
 #include "flat_file.h"
-#include "../../kv_system.h"
+#include "../../key_value/kv_system.h"
 
 err_t
 ff_initialize(
