+++ src/dictionary/open_address_file_hash/CMakeLists.txt
@@ -10,9 +10,9 @@ set(SOURCE_FILES
     ../dictionary.h
     ../dictionary.c
     ../dictionary_types.h
-    ../../kv_io.h
-    ../../kv_io.c
-    ../../kv_system.h)
+        ../../key_value/kv_io.h
+        ../../key_value/kv_io.c
+        ../../key_value/kv_system.h)
 
 if(USE_ARDUINO)
     set(${PROJECT_NAME}_BOARD       ${BOARD})
