+++ src/dictionary/open_address_file_hash/open_address_file_hash_dictionary_handler.c
@@ -7,7 +7,7 @@
 /******************************************************************************/
 
 #include "open_address_file_hash_dictionary_handler.h"
-#include "./../../kv_system.h"
+#include "../../key_value/kv_system.h"
 
 void
 oafdict_init(
