+++ src/dictionary/open_address_hash/open_address_hash_dictionary_handler.h
@@ -15,7 +15,7 @@ extern "C" {
 
 #include "../dictionary_types.h"
 #include "./../dictionary.h"
-#include "./../../kv_system.h"
+#include "../../key_value/kv_system.h"
 #include "open_address_hash.h"
 #include "open_address_hash_dictionary.h"
