+++ src/examples/master_table_example.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include "./../kv_system.h"
+#include "../key_value/kv_system.h"
 #include "./../dictionary/dictionary.h"
 #include "../dictionary/bpp_tree/bpp_tree_handler.h"
 #include "../dictionary/ion_master_table.h"
