+++ src/file/ion_file.h
@@ -14,7 +14,7 @@
 extern "C" {
 #endif
 
-#include "./../kv_system.h"
+#include "../key_value/kv_system.h"
 
 typedef long file_offset_t;
