+++ src/file/linked_file_bag.h
@@ -31,7 +31,7 @@
 extern "C" {
 #endif
 
-#include "./../kv_system.h"
+#include "../key_value/kv_system.h"
 #include "ion_file.h"
 
 #define LFB_NULL FILE_NULL
