+++ src/tests/unit/dictionary/generic_dictionary_test.c
@@ -1,5 +1,5 @@
 #include "generic_dictionary_test.h"
-#include "../../../kv_system.h"
+#include "../../../key_value/kv_system.h"
 
 int
 get_count_index_by_key(
