+++ src/tests/unit/dictionary/generic_dictionary_test.h
@@ -8,7 +8,7 @@ extern "C" {
 #include <stdlib.h>
 #include <limits.h>
 #include "../../planckunit/src/planck_unit.h"
-#include "./../../../kv_system.h"
+#include "../../../key_value/kv_system.h"
 #include "../../../dictionary/dictionary_types.h"
 #include "./../../../dictionary/dictionary.h"
 #include "./../../../dictionary/ion_master_table.h"
